+++ tests/StackExchange.Redis.Tests/Config.cs
@@ -611,4 +611,19 @@ public class Config : TestBase
         var newPass = options.Password = "newPassword";
         Assert.Equal(newPass, conn.RawConfig.Password);
     }
+
+    [Fact]
+    public void HttpTunnel()
+    {
+        var config = ConfigurationOptions.Parse("127.0.0.1:6380,tunnel=http:somewhere:22");
+        var ip = Assert.IsType<IPEndPoint>(Assert.Single(config.EndPoints));
+        Assert.Equal(6380, ip.Port);
+        Assert.Equal("127.0.0.1", ip.Address.ToString());
+
+        Assert.NotNull(config.Tunnel);
+        Assert.Equal("http:somewhere:22", config.Tunnel.ToString());
+
+        var cs = config.ToString();
+        Assert.Equal("127.0.0.1:6380,tunnel=http:somewhere:22", cs);
+    }
 }